requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
spacy
spacy-langdetect
nltk
tensorflow
keras
joblib
ipykernel
opencv-python
scikit-image
pyYAML
python-box